requirements.txt
labelImg==1.8.6
pillow==11.1.0
ultralytics==8.3.59